Fix require statements, move controllers to correct dir
